[#3352] Abstract role synchronization test
